+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Fixed
+- `@rarimo/react-provider`: refactor useProvider hook
+
 ## [1.5.0-rc.5] - 2023-03-30
 ### Added
 - `@rarimo/nft-checkout`: Selectable token to swap